open: Update type inference logic to match getfile
